# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SLZB Spot Monitor",
    "version": "0.2.0",
    "description": "多用户只读现货监控：管理员配置绑定，每个登录用户只读取自己的账户。持久 Cookie 自动续期。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local; replace with your deployment URL"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/session": {
      "get": {
        "operationId": "session",
        "summary": "验证会话",
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "login",
        "summary": "用户名及密码登录（同源 Origin 必需）",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/Origin"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "username": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User info and persistent HttpOnly slzb_session cookie; Max-Age=31536000; rolls forward after one day of activity."
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "logout",
        "summary": "清除会话（同源 Origin 必需）",
        "parameters": [
          {
            "$ref": "#/components/parameters/Origin"
          }
        ],
        "responses": {
          "200": {
            "description": "Cookie cleared"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "slzb_session"
      },
      "cronBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET, separate from API token"
      }
    },
    "parameters": {
      "Origin": {
        "name": "Origin",
        "in": "header",
        "schema": {
          "type": "string",
          "format": "uri"
        },
        "description": "Cookie 写请求必须携带与服务地址一致的 Origin。"
      }
    },
    "responses": {
      "Error": {
        "description": "Sanitized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["error"],
              "properties": {
                "error": {
                  "type": "object",
                  "required": ["code", "message"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}